+++ packages/dts-plugin/src/core/lib/typeScriptCompiler.ts
@@ -1,5 +1,5 @@
 import ansiColors from 'ansi-colors';
-import { dirname, join, normalize, relative, resolve } from 'path';
+import { dirname, join, normalize, relative, resolve, sep } from 'path';
 import typescript from 'typescript';
 import { ThirdPartyExtractor } from '@module-federation/third-party-dts-extractor';
 
@@ -64,7 +64,10 @@ const createHost = (
   const host = typescript.createCompilerHost(tsConfig);
   const originalWriteFile = host.writeFile;
   const mapExposeToEntry = Object.fromEntries(
-    Object.entries(mapComponentsToExpose).map((entry) => entry.reverse()),
+    Object.entries(mapComponentsToExpose).map(([exposed, filename]) => [
+      normalize(filename),
+      exposed,
+    ]),
   );
   const mfTypePath = retrieveMfTypesPath(tsConfig, remoteOptions);
 
@@ -86,7 +89,7 @@ const createHost = (
     );
 
     for (const sourceFile of sourceFiles || []) {
-      const sourceEntry = mapExposeToEntry[sourceFile.fileName];
+      const sourceEntry = mapExposeToEntry[normalize(sourceFile.fileName)];
       if (sourceEntry) {
         const mfeTypeEntry = join(
           mfTypePath,
@@ -95,7 +98,9 @@ const createHost = (
         const mfeTypeEntryDirectory = dirname(mfeTypeEntry);
         const relativePathToOutput = relative(mfeTypeEntryDirectory, filepath)
           .replace(DEFINITION_FILE_EXTENSION, '')
-          .replace(STARTS_WITH_SLASH, '');
+          .replace(STARTS_WITH_SLASH, '')
+          .split(sep) // Windows platform-specific file system path fix
+          .join('/');
         originalWriteFile(
           mfeTypeEntry,
           `export * from './${relativePathToOutput}';\nexport { default } from './${relativePathToOutput}';`,